requirements.txt
pandas~=1.2.2
beautifulsoup4~=4.9.3
imgui[glfw3]~=1.3.0
yfinance~=0.1.55
moderngl-window
zipline
numpy~=1.19.2
bs4~=0.0.1
glfw~=2.0.0
pyrr~=0.10.3
moderngl~=5.6.4
pytz~=2019.3
matplotlib~=3.3.2
yahooquery~=2.2.15
python-binance